refactor: move env vars validation to a separate file
